requirements.txt
streamlit
fpdf2
markdown 